Web Server can now accept connections.
This commit includes the codes for the actual Web Server and the code
for handling connections from clients. Included here is a sample html
page/file for testing/debugging purposes.

Issue: This version can’t find the desired html file and send it to the
requesting client. It can identify the filename from the URL but cannot
load the file itself.

Feel free to make modifications so that it would work perfectly. Thanks!